pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Mourad Abdellahi <medabdellahi1360@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.8"
django = "^5.0.4"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg-binary = "^3.1.17"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
